refactor icons location
